pyproject.toml
[project]
name = "admin-panel-adaptive"
version = "0.1.0"
description = "Adaptive admin panel supporting multiple database backends"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.133.0",
    "jinja2>=3.1.0",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.41.0",
]

[project.optional-dependencies]
mysql = ["pymysql>=1.1.0"]
postgres = ["psycopg2-binary>=2.9.0"]
bigquery = ["sqlalchemy-bigquery>=1.11.0"]
